created posts.js file and added respective code in it
