+++ routes/products.js
@@ -1,7 +1,6 @@
 const express = require("express");
 const router = express.Router();
 const Products = require("../Products.json");
-const Posts = require("../Posts.json");
 
 router.get("/", (req, res) => {
   res.send(Products);
@@ -37,18 +36,6 @@ router.get("/search", (req, res) => {
   }
 });
 
-router.get("/search/:userId", (req, res) => {
-  const userId  = parseInt(req.params.userId);
-  let results = [];
-  Posts.posts.map((post) => {
-    if(post.userId === userId)results.push(post);
-  });
-  if (results.length !== 0) {
-    res.send(results);
-  } else {
-    res.status(404).send("No Product found");
-  }
-  
-})
+
 
 module.exports = router;
